pyproject.toml
[tool.poetry]
name = "sweetdealer - copy"
version = "0.1.0"
description = "SweetDealer backend for React frontend."
authors = ["Henish Patel <200120107182@git.org.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-jazzmin = "^2.6.0"
djangorestframework-simplejwt = "^5.3.0"
django-jsonform = "^2.19.1"
django-cors-headers = "^4.2.0"
docutils = "^0.20.1"
xlsxwriter = "^3.1.7"
django-admin-shortcuts = "^2.1.1"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
